Removes getters/setters, changes in spacing
Refactored code to remove getters and setters after being convinced
by Google that getters and setters are a bad thing. Setting is now
done in the constructor, and instead of getters we have functions
resembling properties.

Also added spacing convention to be 2 blank lines between functions.
This is a convention I stole either from Python or Django, I can't
remember.