Add armor class and hit points
Added armor class and hit points properties to character. This is
used for the charcter to not get hit by an attack, and to take
damage when hit.

Note: both armor class and hit points are hardcoded for now. This
satisfies all the specs that I have encountered so far. I know this
will not be the case as I progress through the rest of the user
stories, but I'm trying to be as YAGNI as possible right now.