refactor to use defaults, rearrange functions
Refactored constructor so that the default arguements are actually
used as defaults. Done because otherwise it was silly to have
default arguments that only existed to later be reassigned.

Also rearranged the order that some of the functions appear, mainly
grouping all the functions pertaining to the character's health
together. This was done just because that's how I felt like grouping
them.